SentimentAnalysis_ModelService/SentimentAPI-Flask-Part3/loadyaml.py
import yaml

def load_yaml(yaml_path):
    try:
        with open(yaml_path, "r") as f:
            config = yaml.safe_load(f)
    except Exception as e:
        logging.error(f"Failed to load a yaml file due to {e}")
    return config